fix: build, response processing; bump version
