--- build.gradle
@@ -1,36 +1,0 @@
-apply plugin: 'java'
-targetCompatibility = 1.8
-sourceCompatibility = 1.8
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    compile 'net.portswigger.burp.extender:burp-extender-api:1.7.13'
-	compile 'com.cedarsoftware:json-io:4.10.1'
-}
-
-sourceSets {
-    main {
-        java {
-            srcDir '.'
-        }
-    }
-}
-
-task fatJar(type: Jar) {
-    baseName = project.name + '-all'
-    from { configurations.compile.collect { it.isDirectory() ? it : zipTree(it) } }
-    with jar
-}
-
-jar {
-    from configurations.compile.collect { it.isDirectory() ? it : zipTree(it) }
-}
-
-tasks.withType(AbstractArchiveTask) {
-    preserveFileTimestamps = false
-    reproducibleFileOrder = true
-}
-
